requirements.txt
requests>=2.31,<3
pycryptodome>=3.20,<4
Pillow>=10,<12
textual>=8.1,<9
prompt_toolkit>=3.0.43,<4
